TIP-1129: Merge into 1 class
